requirements.txt
# Web app
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Google Cloud (search and APIs)
google-cloud-discoveryengine>=0.11.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=0.4.6
google-cloud-storage>=2.14.0
google-cloud-documentai>=2.20.0

# Ingestion helpers
pypdf>=4.0.0
